pregunta2/requirements.txt
flask
mysql-connector-python
awslambdaric
